requirements.txt
Flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
schedule==1.2.1
python-dotenv==1.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1
